+++ lib/tir/defun.ml
@@ -21,7 +21,7 @@ let builtin_names : StringSet.t =
       "int_to_string"; "float_to_string"; "bool_to_string"; "atom_to_string";
       "string_to_int"; "string_length"; "string_concat";
       "string_byte_length"; "string_byte_at"; "string_is_empty";
-      "string_join"; "string_concat3";
+      "string_join"; "string_concat3"; "string_concat_n";
       "read_line"; "io_read_line"; "read_byte"; "io_read_byte"; "not";
       "panic";
       "head"; "tail"; "is_nil";